Move SPI related classes to the SPI package
