+++ extensions/rest/deployment/src/main/java/org/apache/camel/quarkus/component/rest/deployment/RestProcessor.java
@@ -25,8 +25,8 @@ import io.quarkus.deployment.builditem.FeatureBuildItem;
 import org.apache.camel.component.rest.RestComponent;
 import org.apache.camel.quarkus.component.rest.RestRecorder;
 import org.apache.camel.quarkus.component.rest.graal.NoJAXBContext;
-import org.apache.camel.quarkus.core.deployment.CamelServiceFilter;
 import org.apache.camel.quarkus.core.deployment.spi.CamelBeanBuildItem;
+import org.apache.camel.quarkus.core.deployment.spi.CamelServiceFilter;
 import org.apache.camel.quarkus.core.deployment.spi.CamelServiceFilterBuildItem;
 import org.apache.camel.quarkus.support.common.CamelCapabilities;
